Fix VAE decode dtype mismatch: convert latents to vae.dtype before decoding

+++ src/models/my_model.py
@@ -339,13 +339,12 @@ class InsertAnything(L.LightningModule):
         """
         vae = self.flux_fill_pipe.vae
 
-        # Unscale latents
+        latents = latents.to(dtype=vae.dtype)
+
         latents = latents / vae.config.scaling_factor + vae.config.shift_factor
 
-        # Decode
         images = vae.decode(latents).sample
 
-        # Convert from [-1, 1] to [0, 1]
         images = (images + 1) / 2
         images = images.clamp(0, 1)
